quiz functionality is completed

+++ app/src/main/java/com/teamx/equiz/data/models/topWinnerData/Game.kt
@@ -14,6 +14,6 @@ data class Game(
     val observation: Double,
     val rank: Double,
     val speed: Double,
-    val totalScore: Int,
+    val totalScore: Double,
     val userId: UserId
 )
